refactor: DatabaseInitializer 클래스 분리
